+++ watch/cli/prepare_teamfeats.py
@@ -6,6 +6,8 @@ CommandLine:
 
 SeeAlso:
     ~/code/watch/watch/tasks/invariants/predict.py
+    ~/code/watch/watch/tasks/landcover/predict.py
+    ~/code/watch/watch/tasks/depth/predict.py
 
 Example:
     >>> from watch.cli.prepare_teamfeats import *  # NOQA
@@ -95,7 +97,7 @@ Ignore:
 
 
     # Drop 6
-    export CUDA_VISIBLE_DEVICES="1"
+    export CUDA_VISIBLE_DEVICES="0"
     DVC_DATA_DPATH=$(smartwatch_dvc --tags='phase2_data' --hardware=auto)
     BUNDLE_DPATH=$DVC_DATA_DPATH/Drop6
     python -m watch.cli.prepare_teamfeats \
@@ -107,7 +109,7 @@ Ignore:
         --with_depth=0 \
         --do_splits=0 \
         --skip_existing=1 \
-        --gres=1, --workers=1 --backend=serial --run=1
+        --gres=1, --workers=1 --backend=tmux --run=1
 
 
 """